refactor: inner 클래스 정의를 별도 클래스 정의로 변경
