+++ src/main/java/bowling/domain/engine/frame/Score.java
@@ -12,7 +12,7 @@ public class Score {
     private final int value;
     private final int completionCount;
 
-    private Score(int value, int completionCount) {
+    protected Score(int value, int completionCount) {
         this.value = value;
         this.completionCount = completionCount;
     }
@@ -45,32 +45,6 @@ public class Score {
 
     public boolean isCalculationCompleted() {
         return completionCount == 0;
-    }
-
-    public static final class UnavailableScore extends Score {
-
-        protected UnavailableScore(int value, int injectionCount) {
-            super(value, injectionCount);
-        }
-
-        public static UnavailableScore init() {
-            return new UnavailableScore(-1, -1);
-        }
-
-        @Override
-        public Score add(RollResult rollResult) {
-            throw new IllegalStateException("점수 계산이 허용되지 않습니다.");
-        }
-
-        @Override
-        public int getValue() {
-            throw new IllegalStateException("점수 산정을 허용하지 않습니다.");
-        }
-
-        @Override
-        public boolean isCalculationCompleted() {
-            return false;
-        }
     }
 
 }
